split into `backend.rs`/`server.rs`/`wasm.rs`
closes #9
