+++ Cargo.toml
@@ -9,7 +9,11 @@ version = "0.1.0"
 [lib]
 crate-type = ["cdylib", "rlib"]
 name = "turbocharger_generated"
-path = "src/main.rs"
+path = "src/wasm.rs"
+
+[[bin]]
+name = "turbocharger-template"
+path = "src/server.rs"
 
 [dependencies]
 # dependencies for both server and wasm
