+++ src/server.rs
@@ -1,12 +1,7 @@
-use turbocharger::{backend, server_only};
-use wasm_bindgen::prelude::*;
+mod backend;
 
-#[cfg(not(target_arch = "wasm32"))]
 use clap::Clap;
-#[cfg(not(target_arch = "wasm32"))]
-use turbosql::{select, Turbosql};
 
-#[cfg(not(target_arch = "wasm32"))]
 #[derive(Clap)]
 struct Opts {
  #[clap(short, long)]
@@ -17,24 +12,6 @@ struct Opts {
  port: u16,
 }
 
-#[backend]
-#[cfg_attr(not(target_arch = "wasm32"), derive(Turbosql))]
-pub struct Person {
- pub rowid: Option<i64>,
- pub name: Option<String>,
-}
-
-#[backend]
-async fn insert_person(p: Person) -> Result<i64, turbosql::Error> {
- p.insert() // returns rowid
-}
-
-#[backend]
-async fn get_person(rowid: i64) -> Result<Person, turbosql::Error> {
- select!(Person "WHERE rowid = ?", rowid)
-}
-
-#[server_only]
 #[tokio::main]
 async fn main() {
  #[derive(rust_embed::RustEmbed)]
